# docs/api/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SSAC Backend API",
...
    "contact": {
      "name": "SSAC Backend Team",
      "email": "ssac-backend@ssac.com"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/login/oauth2/code/kakao": {
      "get": {
        "tags": [
          "\uce74\uce74\uc624 OAuth"
        ],
        "summary": "\uce74\uce74\uc624 \ub85c\uadf8\uc778 \ucf5c\ubc31",
        "description": "\uce74\uce74\uc624 \uc778\uc99d \ucf54\ub4dc\ub97c \ubc1b\uc544 Access Token\uacfc Refresh Token\uc744 \ubc1c\uae09\ud55c\ub2e4. \ub450 \ud1a0\ud070 \ubaa8\ub450 HttpOnly \ucfe0\ud0a4\ub85c \uc804\ub2ec\ub41c\ub2e4. \uc2e0\uaddc \uc0ac\uc6a9\uc790\ub294 \uc790\ub3d9 \uac00\uc785\ub41c\ub2e4. guestId \ucfe0\ud0a4\uac00 \uc788\uc73c\uba74 \ube44\ud68c\uc6d0 \ud034\uc988 \uae30\ub85d\uc744 \ud68c\uc6d0 \uacc4\uc815\uc73c\ub85c \uc790\ub3d9 \uc774\uc804\ud55c\ub2e4. \uc2e4\uc81c \uc694\uccad\uc740 Spring Security OAuth2 \ud544\ud130\uac00 \ucc98\ub9ac\ud55c\ub2e4.",
        "operationId": "callback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "\uce74\uce74\uc624\uac00 \uc804\ub2ec\ud55c \uc778\uc99d \ucf54\ub4dc",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "CSRF \ubc29\uc5b4\uc6a9 state \ud30c\ub77c\ubbf8\ud130",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "\uc778\uc99d \ud1a0\ud070\uc774 \uc5c6\uac70\ub098 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4. \ud1a0\ud070\uc744 \ud655\uc778\ud558\uc138\uc694.\"}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958. \ub3d9\uc77c\ud55c \uc694\uccad\uc774 \ubc18\ubcf5\ub418\uba74 \ubc31\uc5d4\ub4dc \ud300\uc5d0 \ubb38\uc758\ud558\uc138\uc694.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958\uac00 \ubc1c\uc0dd\ud588\uc2b5\ub2c8\ub2e4.\"}"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/naver/callback": {
      "get": {
        "tags": [
          "\ub124\uc774\ubc84 OAuth"
        ],
        "summary": "\ub124\uc774\ubc84 \ub85c\uadf8\uc778 \ucf5c\ubc31",
        "description": "\ub124\uc774\ubc84 \uc778\uc99d \ucf54\ub4dc\ub97c \ucc98\ub9ac\ud55c\ub2e4. \uae30\uc874 \ud68c\uc6d0\uc774\uba74 authCode\uc640 isNewUser=false\ub85c \ub9ac\ub2e4\uc774\ub809\ud2b8\ud55c\ub2e4. \uc2e0\uaddc \ud68c\uc6d0\uc774\uba74 authCode\uc640 isNewUser=true\ub85c \ud68c\uc6d0 \uac00\uc785 \ud50c\ub85c\uc6b0\ub85c \ub9ac\ub2e4\uc774\ub809\ud2b8\ud55c\ub2e4. JWT/tempToken\uc740 URL\uc5d0 \ub178\ucd9c\ub418\uc9c0 \uc54a\uc73c\uba70 FE\uac00 POST /api/v1/auth/token\uc73c\ub85c \uad50\ud658\ud55c\ub2e4. guestId \ucfe0\ud0a4\uac00 \uc788\uc73c\uba74 \uae30\uc874 \ud68c\uc6d0\uc758 \ube44\ud68c\uc6d0 \ud034\uc988 \uae30\ub85d\uc744 \uc790\ub3d9 \uc774\uc804\ud55c\ub2e4. \ub124\uc774\ubc84\uac00 error \ud30c\ub77c\ubbf8\ud130\ub97c \uc804\ub2ec\ud558\ub294 \uacbd\uc6b0 FE \uc5d0\ub7ec \ud398\uc774\uc9c0\ub85c \ub9ac\ub2e4\uc774\ub809\ud2b8\ud55c\ub2e4.",
        "operationId": "callback_1",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "description": "\ub124\uc774\ubc84\uac00 \uc804\ub2ec\ud55c \uc778\uc99d \ucf54\ub4dc",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "CSRF \ubc29\uc5b4\uc6a9 state \ud30c\ub77c\ubbf8\ud130",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "in": "query",
            "description": "\ub124\uc774\ubc84 \uc5d0\ub7ec \ucf54\ub4dc (\uc0ac\uc6a9\uc790 \uac70\ubd80 \ub4f1)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error_description",
            "in": "query",
            "description": "\ub124\uc774\ubc84 \uc5d0\ub7ec \uc124\uba85",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "guestId",
            "in": "cookie",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "\uc778\uc99d \ud1a0\ud070\uc774 \uc5c6\uac70\ub098 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4. \ud1a0\ud070\uc744 \ud655\uc778\ud558\uc138\uc694.\"}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958. \ub3d9\uc77c\ud55c \uc694\uccad\uc774 \ubc18\ubcf5\ub418\uba74 \ubc31\uc5d4\ub4dc \ud300\uc5d0 \ubb38\uc758\ud558\uc138\uc694.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958\uac00 \ubc1c\uc0dd\ud588\uc2b5\ub2c8\ub2e4.\"}"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/dev/mock-new-user": {
      "get": {
        "tags": [
          "[DEV] \uc778\uc99d \ubaa8\uc758"
        ],
        "summary": "[DEV] \uc2e0\uaddc \ud68c\uc6d0 \uac00\uc785 \ud50c\ub85c\uc6b0 \ubaa8\uc758 \uc2dc\uc791",
        "description": "\uce74\uce74\uc624\u00b7\ub124\uc774\ubc84 OAuth \uc5c6\uc774 tempToken\uc744 \ubc1c\uae09\ud558\uace0 \uc2e0\uaddc \ud68c\uc6d0 \uac00\uc785 \ud50c\ub85c\uc6b0\ub97c \uc2dc\uc791\ud55c\ub2e4. redirect=true(\uae30\ubcf8)\uc774\uba74 FE \ucf5c\ubc31 URL\ub85c \ub9ac\ub2e4\uc774\ub809\ud2b8, redirect=false\uc774\uba74 JSON\uc73c\ub85c tempToken\uc744 \ubc18\ud658\ud55c\ub2e4.",
        "operationId": "mockNewUser",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "KAKAO",
              "enum": [
                "KAKAO",
                "NAVER"
              ]
            }
          },
          {
            "name": "redirect",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "\uc778\uc99d \ud1a0\ud070\uc774 \uc5c6\uac70\ub098 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4. \ud1a0\ud070\uc744 \ud655\uc778\ud558\uc138\uc694.\"}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958. \ub3d9\uc77c\ud55c \uc694\uccad\uc774 \ubc18\ubcf5\ub418\uba74 \ubc31\uc5d4\ub4dc \ud300\uc5d0 \ubb38\uc758\ud558\uc138\uc694.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958\uac00 \ubc1c\uc0dd\ud588\uc2b5\ub2c8\ub2e4.\"}"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/dev/users": {
      "delete": {
        "tags": [
          "[DEV] \uc778\uc99d \ubaa8\uc758"
        ],
        "summary": "[DEV] \ud14c\uc2a4\ud2b8 \uc0ac\uc6a9\uc790 \uc0ad\uc81c",
        "description": "\uc774\uba54\uc77c \uae30\uc900\uc73c\ub85c \uc0ac\uc6a9\uc790 \ubc0f \ubaa8\ub4e0 \uc5f0\uad00 \ub370\uc774\ud130\ub97c \uc0ad\uc81c\ud55c\ub2e4. \ud14c\uc2a4\ud2b8 \ucd08\uae30\ud654 \ubaa9\uc801\uc73c\ub85c \uc0ac\uc6a9\ud55c\ub2e4.",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMapStringString"
                }
              }
            }
          },
          "401": {
            "description": "\uc778\uc99d \ud1a0\ud070\uc774 \uc5c6\uac70\ub098 \ub9cc\ub8cc\ub418\uc5c8\uc2b5\ub2c8\ub2e4.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4. \ud1a0\ud070\uc744 \ud655\uc778\ud558\uc138\uc694.\"}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958. \ub3d9\uc77c\ud55c \uc694\uccad\uc774 \ubc18\ubcf5\ub418\uba74 \ubc31\uc5d4\ub4dc \ud300\uc5d0 \ubb38\uc758\ud558\uc138\uc694.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseError"
                },
                "examples": {
                  "error": {
                    "value": "{\"success\":false,\"data\":null,\"message\":\"\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958\uac00 \ubc1c\uc0dd\ud588\uc2b5\ub2c8\ub2e4.\"}"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiResponseError": {
        "properties": {
          "success": {
            "example": false
          },
          "data": {
            "example": null
          },
          "message": {
            "example": "\uc0ac\uc6a9\uc790\ub97c \ucc3e\uc744 \uc218 \uc5c6\uc2b5\ub2c8\ub2e4."
          }
        }
      },
...
      "ApiResponseMapStringString": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "message": {
            "type": "string"
          },
          "loginRequired": {
            "type": "boolean"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "\ub85c\uadf8\uc778 \ud6c4 \ubc1c\uae09\ubc1b\uc740 JWT \ud1a0\ud070. \ud615\uc2dd: Bearer <token>",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}